roll engine #0a5ab5f1

+++ android/src/main/java/me/yohom/amap_search_fluttify/sub_handler/SubHandler3.java
@@ -2174,7 +2174,7 @@ public class SubHandler3 {
             put("com.amap.api.services.route.TruckPath::setDuration", (args, methodResult) -> {
                 // args
                 // jsonable arg
-                long var1 = (long) ((Map<String, Object>) args).get("var1");
+                int var1 = (int) ((Map<String, Object>) args).get("var1");
             
                 // ref
                 int refId = (int) ((Map<String, Object>) args).get("refId");
@@ -3845,7 +3845,7 @@ public class SubHandler3 {
             put("com.amap.api.services.route.Path::setDuration", (args, methodResult) -> {
                 // args
                 // jsonable arg
-                long var1 = (long) ((Map<String, Object>) args).get("var1");
+                int var1 = (int) ((Map<String, Object>) args).get("var1");
             
                 // ref
                 int refId = (int) ((Map<String, Object>) args).get("refId");
@@ -4915,7 +4915,7 @@ public class SubHandler3 {
             put("com.amap.api.services.nearby.NearbyInfo::setTimeStamp", (args, methodResult) -> {
                 // args
                 // jsonable arg
-                long var1 = (long) ((Map<String, Object>) args).get("var1");
+                int var1 = (int) ((Map<String, Object>) args).get("var1");
             
                 // ref
                 int refId = (int) ((Map<String, Object>) args).get("refId");
